Modifies phone number textview to accept phone input. Also rectifies majority of the issues in bank fragment
