+++ EWRAApp/app/src/main/res/layout/activity_main.xml
@@ -97,7 +97,7 @@
                             android:drawableStart="@drawable/ic_phone_black_24dp"
                             android:drawablePadding="16dp"
                             android:hint="@string/phone_number"
-                            android:inputType="text"
+                            android:inputType="phone"
                             android:maxLines="1"
                             android:textColor="@android:color/black"
                             android:textColorHint="@color/colorPrimary"
