+++ EWRAApp/app/src/main/res/layout/bank_recyclerview_item.xml
@@ -11,7 +11,7 @@
     android:background="#f2f2f2">
 
     <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="400dp"
+        android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:padding="4dp">
