+++ EWRAApp/app/src/main/res/layout/fragment_bank.xml
@@ -1,19 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="8dp"
+    android:layout_marginBottom="4dp"
     tools:context=".ui.bank.BankFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_bankFragment"
-            android:layout_width="409dp"
-            android:layout_height="729dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
             tools:listitem="@layout/bank_recyclerview_item" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+</FrameLayout>
